- Changed so that error message appears before adding media if on playlist selected - Reorganising dependency properties - Removed separate class for dependencyproperties, it can not be created in a separate class

+++ ProgrammeringCSharp3/MultiMediaClassesAndManagers/MediaSubClasses/Video.cs
@@ -10,7 +10,7 @@ namespace MultiMediaClassesAndManagers.MediaSubClasses
     /// </summary>
     public class Video : MediaFile
     {
-        public int LengthInMinutes { get; set; }
+        public double LengthInMinutes { get; set; }
 
         /// <summary>
         /// The Video class constructor, initializes the object of type Video (as well as passes on the name and other common ground information to the base class)
@@ -20,7 +20,7 @@ namespace MultiMediaClassesAndManagers.MediaSubClasses
         /// <param name="previewUrl">Url of preview (differs for type)</param>
         /// <param name="fileExtention">The fileExtention of the video</param>
         /// <param name="lengthInMinutes">The vides length</param>
-        public Video(string name, string source, string previewUrl, string fileExtention, int lengthInMinutes) : base(name, source, previewUrl, fileExtention)
+        public Video(string name, string source, string previewUrl, string fileExtention, double lengthInMinutes) : base(name, source, previewUrl, fileExtention)
         {
             LengthInMinutes = lengthInMinutes;
         }
